src/requirements.txt
fastapi==0.104.1
pydantic==2.5.2
PyYAML==6.0.1
uvicorn==0.24.0
python-gitlab==4.2.0